+++ assets/src/edit-story/components/dropzone/dropZoneProvider.js
@@ -28,7 +28,10 @@ import { useCallback, useState } from '@wordpress/element';
 /**
  * Internal dependencies
  */
+import { useUploader } from '../../app/uploader';
+import { useMedia } from '../../app/media';
 import Context from './context';
+import getDragType from './utils/getDragType';
 
 const DropZoneWrapper = styled.div`
   height: 100%;
@@ -37,6 +40,11 @@ const DropZoneWrapper = styled.div`
 function DropZoneProvider({ children }) {
   const [dropZones, setDropZones] = useState([]);
   const [hoveredDropZone, setHoveredDropZone] = useState(null);
+  const { uploadFile } = useUploader();
+  const {
+    state: { DEFAULT_WIDTH },
+    actions: { insertMediaElement },
+  } = useMedia();
 
   const registerDropZone = useCallback(
     (dropZone) => {
@@ -76,7 +84,7 @@ function DropZoneProvider({ children }) {
   };
 
   const onDragOver = (evt) => {
-    evt.preventDefault();
+    disableDefaults(evt);
     // Get the hovered dropzone. // @todo Consider dropzone inside dropzone, will we need this?
     const foundDropZones = dropZones.filter((dropZone) => {
       return isWithinElementBounds(dropZone.node, evt.clientX, evt.clientY);
@@ -114,6 +122,41 @@ function DropZoneProvider({ children }) {
     }
   };
 
+  const disableDefaults = (evt) => {
+    evt.preventDefault();
+    evt.stopPropagation();
+  };
+
+  const onDrop = (evt) => {
+    disableDefaults(evt);
+    if ('file' === getDragType(evt)) {
+      const files = [...evt.dataTransfer.files];
+      files.forEach((file) => {
+        uploadFile(file).then(
+          ({
+            id,
+            guid: { rendered: src },
+            media_details: { width: oWidth, height: oHeight },
+            mime_type: mimeType,
+            featured_media: posterId,
+            featured_media_src: poster,
+          }) => {
+            const mediaEl = {
+              id,
+              posterId,
+              poster,
+              src,
+              oWidth,
+              oHeight,
+              mimeType,
+            };
+            insertMediaElement(mediaEl, DEFAULT_WIDTH, false);
+          }
+        );
+      });
+    }
+  };
+
   const state = {
     state: {
       hoveredDropZone,
@@ -126,7 +169,12 @@ function DropZoneProvider({ children }) {
     },
   };
   return (
-    <DropZoneWrapper onDragOver={onDragOver}>
+    <DropZoneWrapper
+      onDragOver={onDragOver}
+      onDragLeave={disableDefaults}
+      onDragEnter={disableDefaults}
+      onDrop={onDrop}
+    >
       <Context.Provider value={state}>{children}</Context.Provider>
     </DropZoneWrapper>
   );
