+++ assets/src/edit-story/components/library/libraryContent.js
@@ -17,13 +17,7 @@
 /**
  * Internal dependencies
  */
-import {
-  DEFAULT_EDITOR_PAGE_WIDTH,
-  DEFAULT_EDITOR_PAGE_HEIGHT,
-} from '../../constants';
-import { createNewElement, getDefinitionForType } from '../../elements';
-import { editorToDataX, editorToDataY } from '../../units';
-import { useStory } from '../../app';
+import useInsertElement from '../../elements/shared/useInsertElement';
 import useLibrary from './useLibrary';
 import MediaLibrary from './mediaLibrary';
 import TextLibrary from './textLibrary';
@@ -37,38 +31,17 @@ function Library() {
       tabs: { MEDIA, TEXT, SHAPES, LINKS },
     },
   } = useLibrary();
-  const {
-    actions: { addElement, setBackgroundElement },
-    state: { currentPage },
-  } = useStory();
+
+  const insertElement = useInsertElement();
+
   const ContentLibrary = {
     [MEDIA]: MediaLibrary,
     [TEXT]: TextLibrary,
     [SHAPES]: ShapeLibrary,
     [LINKS]: LinkLibrary,
   }[tab];
-  const isMediaEl = (type) => {
-    const { isMedia } = getDefinitionForType(type);
-    return isMedia;
-  };
-  const handleInsert = (type, { width, height, ...props }) => {
-    const element = createNewElement(type, {
-      ...props,
-      x: editorToDataX(80 * Math.random(), DEFAULT_EDITOR_PAGE_WIDTH),
-      y: editorToDataY(70 * Math.random(), DEFAULT_EDITOR_PAGE_HEIGHT),
-      width: editorToDataX(width, DEFAULT_EDITOR_PAGE_WIDTH),
-      height: editorToDataY(height, DEFAULT_EDITOR_PAGE_HEIGHT),
-    });
-    addElement({ element });
-    if (
-      isMediaEl(type) &&
-      !currentPage.elements.some(({ type: elType }) => isMediaEl(elType))
-    ) {
-      setBackgroundElement({ elementId: element.id });
-    }
-    return element;
-  };
-  return <ContentLibrary onInsert={handleInsert} />;
+
+  return <ContentLibrary onInsert={insertElement} />;
 }
 
 export default Library;
