+++ assets/src/edit-story/elements/square/index.js
@@ -25,6 +25,7 @@ export { default as LayerIcon } from './icon';
 
 export const defaultAttributes = {
   backgroundColor: '#ffffff',
+  isBackground: false,
 };
 
 export const hasEditMode = false;
