+++ assets/src/edit-story/elements/text/index.js
@@ -37,6 +37,7 @@ export const defaultAttributes = {
   letterSpacing: 'normal',
   lineHeight: 1.3,
   textAlign: 'initial',
+  isBackground: false,
 };
 
 export const hasEditMode = true;
