+++ assets/src/edit-story/elements/video/index.js
@@ -32,6 +32,7 @@ export const defaultAttributes = {
   posterId: null,
   poster: null,
   videoId: 0,
+  isBackground: true,
 };
 
 export const hasEditMode = true;
